remove rename of dir, rename test
