+++ tests/functional/deprecations/fixtures.py
